% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{check_M}
\alias{check_M}
\alias{checkM}
\alias{assertM}
\title{Check if a vector of values is a vector of unique modes}
\usage{
checkM(...)

assertM(...)
}
\description{
Check if a vector of values is a vector of unique modes
}
\section{Functions}{
\itemize{
\item \code{checkM()}: check the vector M

\item \code{assertM()}: assert the vector M

}}
\keyword{internal}
