% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{check_probability}
\alias{check_probability}
\alias{checkProbabilityVec}
\alias{assertProbabilityVec}
\alias{checkProbabilityScalar}
\alias{assertProbabilityScalar}
\title{Check if a vector is a vector of probability}
\usage{
checkProbabilityVec(
  x,
  len = NULL,
  min.len = 1L,
  max.len = NULL,
  unique = FALSE,
  striclyPos = FALSE,
  striclyUnsure = FALSE,
  any.missing = FALSE
)

assertProbabilityVec(
  x,
  len = NULL,
  min.len = 1L,
  max.len = NULL,
  unique = FALSE,
  striclyPos = FALSE,
  striclyUnsure = FALSE,
  any.missing = FALSE,
  .var.name = checkmate::vname(x),
  add = NULL
)

checkProbabilityScalar(...)

assertProbabilityScalar(...)
}
\description{
Check if a vector is a vector of probability
}
\section{Functions}{
\itemize{
\item \code{checkProbabilityVec()}: check a vector of probability

\item \code{assertProbabilityVec()}: assert a vector of probability

\item \code{checkProbabilityScalar()}: check a scalar of probability

\item \code{assertProbabilityScalar()}: assert a scalar of probability

}}
\keyword{internal}
