% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{check_variances}
\alias{check_variances}
\alias{checkVariances}
\alias{assertVariances}
\alias{checkCovarMat}
\alias{assertCovarMat}
\alias{checkCorMat}
\alias{assertCorMat}
\title{Check related to variances}
\usage{
checkVariances(...)

assertVariances(...)

checkCovarMat(mat, p = nrow(mat))

assertCovarMat(
  mat,
  p = nrow(mat),
  .var.name = checkmate::vname(mat),
  add = NULL
)

checkCorMat(mat, p = nrow(mat))

assertCorMat(mat, p = nrow(mat), .var.name = checkmate::vname(mat), add = NULL)
}
\description{
Check related to variances
}
\section{Functions}{
\itemize{
\item \code{checkVariances()}: Check variances and standard deviations

\item \code{assertVariances()}: Assert variances and standard deviations

\item \code{checkCovarMat()}: Check a variance-covariance matrix

\item \code{assertCovarMat()}: Assert a variance-covariance matrix

\item \code{checkCorMat()}: Check a correlation matrix

\item \code{assertCorMat()}: Assert a correlation matrix

}}
\keyword{internal}
