% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{checkPhi_tab}
\alias{checkPhi_tab}
\title{Check if the ponderation for each unit is a convex weight}
\usage{
checkPhi_tab(phi_tab, N = nrow(phi_tab), M)
}
\arguments{
\item{N}{number of statistical units. (stricly positive integer)}

\item{M}{list of expected modes. Must be non-empty, composed of strings
and has to be made of unique values. (character vector)}
}
\description{
Check if the ponderation for each unit is a convex weight
}
\keyword{internal}
