% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{checkX}
\alias{checkX}
\title{Check covariates}
\usage{
checkX(X, N = nrow(X), p = ncol(X))
}
\arguments{
\item{N}{number of statistical units. (stricly positive integer)}
}
\description{
Check covariates
}
\keyword{internal}
