% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{checkY_tab}
\alias{checkY_tab}
\title{Check the table of outcomes.}
\usage{
checkY_tab(Y_tab, N = nrow(Y_tab), K = ncol(Y_tab), M = colnames(Y_tab))
}
\arguments{
\item{Y_tab}{table containing outcome values for the different modes
available. May contain missing values. Available values must be finite.}

\item{N}{number of statistical units. (stricly positive integer)}

\item{M}{list of expected modes. Must be non-empty, composed of strings
and has to be made of unique values. (character vector)}
}
\description{
One row is a unit and one column a mode
}
\keyword{internal}
