% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_tot_estim.R
\name{estim_appr_var_seq_phi1}
\alias{estim_appr_var_seq_phi1}
\title{Estimates the approximate variance of the HT estimator of t_phi1 with known
or estimated mode selection probabilities. Homoscedasticity and independence
in the m1 mode selection mechanism assumed.}
\usage{
estim_appr_var_seq_phi1(
  Yobs,
  modes,
  I,
  pi_mat,
  p1,
  sd1 = NULL,
  phi = rep(1, length(Yobs)),
  correcEstimWeights = FALSE,
  Z = matrix(1, nrow = len(Yobs), ncol = 1L)
)
}
\description{
Estimates the approximate variance of the HT estimator of t_phi1 with known
or estimated mode selection probabilities. Homoscedasticity and independence
in the m1 mode selection mechanism assumed.
}
