% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{estim_appr_var_seq_phi2}
\alias{estim_appr_var_seq_phi2}
\title{Estimates the approximate variance of the HT estimator of t_phi2 with known
or estimated mode selection probabilities. Homoscedasticity is assumed.}
\usage{
estim_appr_var_seq_phi2(
  Yobs,
  modes,
  I,
  piMat,
  pq1Mat,
  pq2Mat,
  Z,
  phi = rep(1, length(Yobs)),
  sd2 = 0,
  correcEstimWeights = FALSE,
  independenceq1 = NULL,
  independenceq2 = NULL,
  ...
)
}
\description{
Estimates the approximate variance of the HT estimator of t_phi2 with known
or estimated mode selection probabilities. Homoscedasticity is assumed.
}
