% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{estim_counterfactuals}
\alias{estim_counterfactuals}
\title{Estimation of the counterfactuals of each respondent}
\usage{
estim_counterfactuals(Yobs, modes, X, ...)
}
\arguments{
\item{Yobs}{vector of the observed outcomes.
(numeric vector of size N the size of the population).}

\item{modes}{vector of the selected mode of each unit.
The first mode of the protocol is defined as "m1", the second as "m2"
(character vector or factor of size N).}

\item{X}{design matrix. The rows corresponding to the non-respondents
can contain NA (numeric matrix with N rows and p columns).}

\item{...}{arguments for the function \code{MatchIt::matchit}.}
}
\value{
for each mode a vector of size N containing for each respondent
its outcome or its estimated counterfactual depending on if it answered
by the mode or the other. Equal to NA for the non-respondents
(numeric matrix of dimension (N,2)).
}
\description{
In this version considers only the case with two modes. Could be
generalised. Requires the package \href{https://cran.r-project.org/web/packages/MatchIt/}{MatchIt}.
}
\seealso{
\code{\link[MatchIt:matchit]{MatchIt::matchit()}}
}
