% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{estim_MC}
\alias{estim_MC}
\alias{bias_MC}
\alias{var_MC}
\alias{sd_MC}
\alias{MSE_MC}
\alias{CV_MC}
\title{Évaluation des estimations via Monte-Carlo}
\usage{
bias_MC(estims, target)

var_MC(estims)

sd_MC(estims)

MSE_MC(estims, target)

CV_MC(estims, target = mean(estims))
}
\arguments{
\item{estims}{vecteur numérique des estimations}

\item{target}{valeur cible / vraie valeur (réelle)}
}
\description{
Évaluation des estimations via Monte-Carlo
}
\section{Functions}{
\itemize{
\item \code{bias_MC()}: Estimation du biais

\item \code{var_MC()}: Variance empirique corrigée des estimations

\item \code{sd_MC()}: Écart-type empirique corrigé des estimations

\item \code{MSE_MC()}: Erreur quadratique moyenne par rapport à l'estimateur

\item \code{CV_MC()}: Coefficient de variation (par défaut estimé via la
moyenne des estimations)

}}
