% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_tot_seq.R
\name{estim_t0}
\alias{estim_t0}
\alias{estim_tot_0_no_p1}
\alias{estim_tot_a}
\title{Estimation of t0}
\usage{
estim_tot_0_no_p1(Yobs, modes, pi, p2)

estim_tot_a(Yobs, responses, pi, pa)
}
\arguments{
\item{Yobs}{vector of the observed outcomes.
(numeric vector of size N the size of the population).}

\item{modes}{vector of the selected mode of each unit.
The first mode of the protocol is defined as "m1", the second as "m2"
(character vector or factor of size N).}

\item{pi}{vector containing the first-order inclusion probabilities
(numeric vector of size N).}

\item{p2}{vector containing the true or estimated selection probabilities
of the second mode (numeric vector of size N).}

\item{responses}{vector indicating which unit answered.
If a unit answered then the corresponding value is "r"
(character vector or factor of size N).}

\item{pa}{vector containing the true or estimated global answer probability
(numeric vector of size N).}
}
\value{
the estimation of the total \eqn{t_0} (scalar).
}
\description{
Estimation of the total \eqn{t_0} under an almost sure equality
between the potential outcomes of each unit. The total can be estimated
using the mode selection probabilities or the global answer probabilities.
}
\section{Functions}{
\itemize{
\item \code{estim_tot_0_no_p1()}: Sum an estimator per mode. Do not use the first mode
selection probabilities.

\item \code{estim_tot_a()}: Use the probability to answer by any mode instead
of by each mode.

}}
\seealso{
Other total estimators: 
\code{\link{estim_tot_m1}()},
\code{\link{estim_tot_m2}()},
\code{\link{estim_tot_phi_y}()}
}
\concept{total estimators}
