% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_tot_multimode.R
\name{estim_tot_a}
\alias{estim_tot_a}
\title{Estimation of t_ay}
\usage{
estim_tot_a(Yobs, responses, pi, pa)
}
\arguments{
\item{Yobs}{vector of the observed outcomes. For the non-respondents
the value is not considered and therefore can be equal to NA
(numeric vector of size N the size of the population).}

\item{responses}{vector indicating which unit answered.
If a unit answered then the corresponding value is "r"
(character vector or factor of size N).}

\item{pi}{vector containing the inclusion probabilities pi_k. Can be
equal to NA for the non-respondents (numeric vector of size N).}

\item{pa}{vector containing the true or estimated probabilities p_ak.
Can be equal to NA for the non-respondents (numeric vector of size N).}
}
\value{
the estimation of the total t_ay (scalar).
}
\description{
Estimation of t_ay
}
