% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_tot_seq.R
\name{estim_tot_m2}
\alias{estim_tot_m2}
\title{Estimation of t2}
\usage{
estim_tot_m2(Yobs, modes, pi, p1, p2, phi = rep(1, length(Yobs)))
}
\arguments{
\item{Yobs}{vector of the observed outcomes.
(numeric vector of size N the size of the population).}

\item{modes}{vector of the selected mode of each unit.
The first mode of the protocol is defined as "m1", the second as "m2"
(character vector or factor of size N).}

\item{pi}{vector containing the first-order inclusion probabilities
(numeric vector of size N).}

\item{p1}{vector containing the true or estimated selection probabilities
of the first mode (numeric vector of size N).}

\item{p2}{vector containing the true or estimated selection probabilities
of the second mode (numeric vector of size N).}

\item{phi}{optional vector containing weights for the total.
The values must be between 0 and 1 (numeric vector size N).}
}
\value{
the estimation of the total \eqn{t_2}, possibly weighted (scalar).
}
\description{
Estimation of the total of the second mode of response
potential outcomes \eqn{t_2}.
}
\seealso{
Other total estimators: 
\code{\link{estim_t0}},
\code{\link{estim_tot_m1}()},
\code{\link{estim_tot_phi_y}()}
}
\concept{total estimators}
