% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_selection_prob.R
\name{Fisher_Information_Matrix}
\alias{Fisher_Information_Matrix}
\title{Fisher Information Matrix (FIM) for a logistic model and weights
equal to 1}
\usage{
Fisher_Information_Matrix(prob, Z, maskSubset = !logical(nrow(Z)))
}
\arguments{
\item{prob}{probability to answer. Can be true or estimated.
Can be equal to NA for the units that are not in the subset defined
by \code{maskSubset} (numeric vector of size N the size of the population).}

\item{Z}{design matrix. The rows corresponding to the units that are not in
the subset can contain NA (numeric matrix with N rows and q columns).}

\item{maskSubset}{mask indicating which unit should be used
in the calculation of the FIM. Default to the entire set
(logical vector of size N).}
}
\value{
the Fisher Information Matrix or its estimation
(numeric matrix of order q).
}
\description{
Fisher Information Matrix (FIM) for a logistic model and weights
equal to 1
}
