% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMsampling-package2.R
\docType{package}
\name{MMsampling-package}
\alias{MMsampling}
\alias{MMsampling-package}
\title{MMsampling: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://atero18.github.io/MMsampling/}
}

}
\author{
\strong{Maintainer}: Tangi Tassin \email{tangi.tassin@outlook.fr} (\href{https://orcid.org/0009-0004-5956-2910}{ORCID})

}
\keyword{internal}
