% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{param_I}
\alias{param_I}
\title{Definition of the sample indicator vector \code{I}}
\arguments{
\item{I}{vector of boolean indicating for each unit if it has been selected
(TRUE) or not (FALSE) in the sample \code{S}.}
}
\description{
Definition of the sample indicator vector \code{I}
}
\keyword{internal}
