% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{param_M}
\alias{param_M}
\title{Definition of the mode list \code{M}}
\arguments{
\item{M}{list of expected modes. Must be non-empty, composed of strings
and has to be made of unique values. (character vector)}
}
\description{
Definition of the mode list \code{M}
}
\keyword{internal}
