% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{param_N}
\alias{param_N}
\title{Definition of the number of units \code{N}}
\arguments{
\item{N}{number of statistical units. (stricly positive integer)}
}
\description{
Definition of the number of units \code{N}
}
\keyword{internal}
