% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{pi2_to_covarInc}
\alias{pi2_to_covarInc}
\title{Calculation of the inclusion covariance matrix using the
second order inclusion probabilities.}
\usage{
pi2_to_covarInc(pi_mat)
}
\arguments{
\item{pi_mat}{matrix containing the second-order inclusion
probabilities pi_kl (symmetric numeric matrix of order N).}
}
\value{
the corresponding covariance matrix
(symmetric numeric matrix of order N).
}
\description{
Calculation of the inclusion covariance matrix using the
second order inclusion probabilities.
}
