% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_choice_prob.R
\name{proba_mode_estim}
\alias{proba_mode_estim}
\alias{estim_response_prob_global}
\alias{estim_response_prob_sequential}
\title{Estimation of mode choising probabilities}
\usage{
estim_response_prob_global(
  sample,
  RGH = NULL,
  constRGH = TRUE,
  chosenOnly = FALSE
)

estim_response_prob_sequential(
  X,
  modes,
  orderModes,
  RGH = NULL,
  constRGH = TRUE,
  link = "logit",
  chosenOnly = FALSE
)
}
\arguments{
\item{sample}{object containing data about a sample and its context}

\item{orderModes}{the order modes are considered. The first mode will have
an absolute probability of selection, the second a probability a selection
conditionaly to the non-selection of the first one, etc.}
}
\description{
Estimation of mode choising probabilities
}
\section{Functions}{
\itemize{
\item \code{estim_response_prob_global()}: Estimation by a multimodal point of view

\item \code{estim_response_prob_sequential()}: Estimation sequentialy with conditional
probabilities (interesting particularly with sequential mixed-modes protocoles)

}}
\keyword{internal}
