% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_measure_bias.R
\name{regression_Ym}
\alias{regression_Ym}
\alias{MCO_Ym}
\title{Parametric regression for modes}
\usage{
MCO_Ym(Xm, Ym, m = NULL, choixMode = NULL)
}
\description{
Estimate Ym on the whole population based
on the subpopulation that used this mode
to answer
}
\keyword{internal}
