% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampling_mode}
\alias{sampling_mode}
\alias{sim_ind_MM}
\alias{sim_ind_MM_eq_prob}
\title{Simulation du choix des modes sous un modèle de Poisson}
\usage{
sim_ind_MM(p, seed = NULL)

sim_ind_MM_eq_prob(N, modes, addNR = TRUE)
}
\arguments{
\item{seed}{graine aléatoire à utiliser pour la reproductibilité. Si vaut
\code{NULL} alors la graine n'est pas fixée.}

\item{N}{Taille de la population}

\item{modes}{Liste des différents modes disponibles (vecteur). Possibilité
que ce soit un entier, auquel cas on considérera autant de modes, plus
éventuellement de la non-réponse.}
}
\description{
Simulation du choix des modes sous un modèle de Poisson
}
\section{Functions}{
\itemize{
\item \code{sim_ind_MM()}: Simulation d'un choix de mode pour chaque
individu de manière indépendante

\item \code{sim_ind_MM_eq_prob()}: Simulation de sélection avec probabilités égales
pour tous les modes et tous les individus

}}
