% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{seed}
\alias{seed}
\alias{fix_seed}
\title{Simulation parameter}
\usage{
fix_seed(seed = NULL)
}
\arguments{
\item{seed}{graine aléatoire à utiliser pour la reproductibilité. Si vaut
\code{NULL} alors la graine n'est pas fixée.}
}
\description{
Simulation parameter
}
\keyword{internal}
