% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{set_I}
\alias{set_I}
\title{Assert the \code{I} vector
Assert if the indicator vector \code{I} is correct, and return it as a boolean
vector if it is made of integers}
\usage{
set_I(I, N)
}
\arguments{
\item{I}{vector of boolean indicating for each unit if it has been selected
(TRUE) or not (FALSE) in the sample \code{S}.}

\item{N}{number of statistical units. (stricly positive integer)}
}
\description{
Assert the \code{I} vector
Assert if the indicator vector \code{I} is correct, and return it as a boolean
vector if it is made of integers
}
\keyword{internal}
