% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_tot_estim.R
\name{var_expansion_seq_phi1}
\alias{var_expansion_seq_phi1}
\title{Calculate the true variance of the HT estimator of t_phi1 in the case
of known selection probabilities. Using estimated values give an estimation
The expected values of the m1 counterfactuals have to be given.
The calculation will be correct under the additional strong assumption
of conditional independence between the sampling design p
and the mode selection mechanism q1.}
\usage{
var_expansion_seq_phi1(Y1exp, pi_mat, p1, phi = rep(1, length(Y1exp)), sd1 = 0)
}
\description{
Calculate the true variance of the HT estimator of t_phi1 in the case
of known selection probabilities. Using estimated values give an estimation
The expected values of the m1 counterfactuals have to be given.
The calculation will be correct under the additional strong assumption
of conditional independence between the sampling design p
and the mode selection mechanism q1.
}
