% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_tot_estim.R
\name{var_expansion_seq_phi2}
\alias{var_expansion_seq_phi2}
\title{Calculate the true variance of the HT estimator of t_phi2 in the case
of known selection probabilities. Using estimated values give an estimation
The expected values of the m2 counterfactuals have to be given.
The calculation will be correct under the additional strong assumptions
of conditional independence between the sampling design p
and the mode selection mechanism q1 and q2, plus the conditional independence
between the mode selection mechanisms.}
\usage{
var_expansion_seq_phi2(
  Y2exp,
  pi_mat,
  p1,
  p2,
  sd2 = 0,
  phi = rep(1, length(Y2exp))
)
}
\description{
Calculate the true variance of the HT estimator of t_phi2 in the case
of known selection probabilities. Using estimated values give an estimation
The expected values of the m2 counterfactuals have to be given.
The calculation will be correct under the additional strong assumptions
of conditional independence between the sampling design p
and the mode selection mechanism q1 and q2, plus the conditional independence
between the mode selection mechanisms.
}
