% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{var_expansion_seq_phi2}
\alias{var_expansion_seq_phi2}
\title{Calculate the true variance of the HT estimator of t_phi2
(or t_2 as a special case) in the case of known selection probabilities.
Using estimated values give an approximate variance in the case of
known selection probabilities.}
\usage{
var_expansion_seq_phi2(
  Y2exp,
  I,
  pi_mat,
  p1_mat,
  p2_mat,
  Z,
  biasedMode,
  refMode,
  modes,
  phi = rep(1, length(Y2exp)),
  correcEstimWeights = FALSE,
  constY2 = FALSE,
  covarY2 = NULL
)
}
\description{
Calculate the true variance of the HT estimator of t_phi2
(or t_2 as a special case) in the case of known selection probabilities.
Using estimated values give an approximate variance in the case of
known selection probabilities.
}
