% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{estim_appr_var_seq_phi1}
\alias{estim_appr_var_seq_phi1}
\title{Estimates the approximate variance of the HT estimator of t_phi1 with known
or estimated mode selection probabilities.}
\usage{
estim_appr_var_seq_phi1(
  Yobs,
  modes,
  I,
  piMat,
  pq1Mat,
  Z,
  phi = rep(1, length(Yobs)),
  correcEstimWeights = FALSE
)
}
\description{
Estimates the approximate variance of the HT estimator of t_phi1 with known
or estimated mode selection probabilities.
}
