% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_tot_multimode.R
\name{estim_tot_m2}
\alias{estim_tot_m2}
\title{Estimation of t_2}
\usage{
estim_tot_m2(Yobs, modes, pi, p1, p2, phi = rep(1, length(Yobs)))
}
\arguments{
\item{Yobs}{vector of the observed outcomes. For the m2 non-respondents
the value is not considered and therefore can be equal to NA
(numeric vector of size N the size of the population).}

\item{modes}{vector of the selected mode of each unit.
The second mode of the protocol is defined as "m2"
(character vector or factor of size N).}

\item{pi}{vector containing the inclusion probabilities pi_k. Can be
equal to NA for the m2 non-respondents (numeric vector of size N).}

\item{p1}{vector containing the true or estimated probabilities p_1k.
Can be equal to NA for the m2 non-respondents (numeric vector of size N).}

\item{p2}{vector containing the true or estimated probabilities p_2k.
Can be equal to NA for the m2 non-respondents (numeric vector of size N).}

\item{phi}{optional vector containing weights for the total. Must be
known for the m2-respondents (numeric vector of \link{0,1}^N).}
}
\value{
the estimation of the total t_2, possibly weighted (scalar).
}
\description{
Estimation of t_2
}
