% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{mats_interaction}
\alias{mats_interaction}
\alias{cor2cov}
\title{Transformation de matrices d'interaction}
\usage{
cor2cov(cor, sigmas)
}
\arguments{
\item{cor}{Matrice de corrélation (matrice)}

\item{sigmas}{Vecteur des écarts-types. Doit être de même taille
que l'ordre de \code{cor}.}
}
\description{
Les fonctions suivantes transforment des matrices contenant des
informations sur les interactions entre variables (ex : corrélation)
en d'autres matrices.
}
\section{Functions}{
\itemize{
\item \code{cor2cov()}: Transformation d'une matrice de corrélation
en une matrice de variance.

}}
\keyword{internal}
