% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_selection_prob.R
\name{proba_mode_estim}
\alias{proba_mode_estim}
\alias{estim_response_prob_global}
\alias{estim_response_prob_sequential}
\title{Estimation of mode choising probabilities}
\usage{
estim_response_prob_global(
  I,
  modes,
  Z,
  RGH = NULL,
  constRGH = TRUE,
  chosenOnly = FALSE
)

estim_response_prob_sequential(
  I,
  Z,
  modes,
  orderModes,
  RGH = NULL,
  constRGH = FALSE,
  link = "logit",
  chosenOnly = TRUE
)
}
\arguments{
\item{orderModes}{the order modes are considered. The first mode will have
an absolute probability of selection, the second a probability a selection
conditionally to the non-selection of the first one, etc.}
}
\description{
Estimation of mode choising probabilities
}
\section{Functions}{
\itemize{
\item \code{estim_response_prob_global()}: Estimation by a multimodal point of view

\item \code{estim_response_prob_sequential()}: Estimation sequentially with conditional
probabilities (interesting particularly with sequential mixed-mode protocols)

}}
\keyword{internal}
