% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{var_expansion_seq_phi1}
\alias{var_expansion_seq_phi1}
\title{Calculate the true variance of the HT estimator of t_phi1 in the case
of known selection probabilities. Using estimated values give an approximate
variance.}
\usage{
var_expansion_seq_phi1(
  Y1exp,
  I,
  pi_mat,
  p1_mat,
  Z,
  biasedMode,
  modes,
  phi = rep(1, length(Y1exp)),
  correcEstimWeights = FALSE,
  constY1 = FALSE,
  covarY1 = NULL
)
}
\description{
Calculate the true variance of the HT estimator of t_phi1 in the case
of known selection probabilities. Using estimated values give an approximate
variance.
}
