% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_tot_multimode.R
\name{var_HT_seq_phi2}
\alias{var_HT_seq_phi2}
\title{Calculate the true variance of the HT estimator of t_phi2
(or t_2 as a special case) in the case of known selection probabilities.
Using estimated values give an approximate variance in the case of
known selection probabilities.}
\usage{
var_HT_seq_phi2(
  expY2,
  covarY2,
  piMat,
  pq1Mat,
  pq2Mat,
  phi = numeric(length(expY2))
)
}
\description{
Calculate the true variance of the HT estimator of t_phi2
(or t_2 as a special case) in the case of known selection probabilities.
Using estimated values give an approximate variance in the case of
known selection probabilities.
}
