% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_tot_multimode.R
\name{var_HT_seq_phi2}
\alias{var_HT_seq_phi2}
\title{Calculate the true variance of the HT estimator of t_phi2
(or t_2 as a special case) in the case of known selection probabilities.
Using estimated values give an approximate variance in the case of
known selection probabilities.}
\usage{
var_HT_seq_phi2(
  expY2,
  I,
  piMat,
  pq1Mat,
  pq2Mat,
  Z,
  biasedMode,
  refMode,
  modes,
  phi = rep(1, length(expY2)),
  correcEstimWeights = FALSE,
  constY2 = FALSE,
  covarY2 = NULL
)
}
\description{
Calculate the true variance of the HT estimator of t_phi2
(or t_2 as a special case) in the case of known selection probabilities.
Using estimated values give an approximate variance in the case of
known selection probabilities.
}
